RSDK-6963: Pull gRPC handling of video streams from gostream up to a higher level package.
