+++ robot/web/web_c.go
@@ -35,7 +35,7 @@ import (
 // StreamServer manages streams and displays.
 type StreamServer struct {
 	// Server serves streams
-	Server gostream.StreamServer
+	Server *webstream.Server
 	// HasStreams is true if service has streams that require a WebRTC connection.
 	HasStreams bool
 }
@@ -107,7 +107,7 @@ func (svc *webService) addNewStreams(ctx context.Context) error {
 		stream, err := svc.streamServer.Server.NewStream(config)
 
 		// Skip if stream is already registered, otherwise raise any other errors
-		var registeredError *gostream.StreamAlreadyRegisteredError
+		var registeredError *webstream.StreamAlreadyRegisteredError
 		if errors.As(err, &registeredError) {
 			return nil, true, nil
 		} else if err != nil {
@@ -155,7 +155,7 @@ func (svc *webService) makeStreamServer(ctx context.Context) (*StreamServer, err
 		if len(svc.videoSources) != 0 || len(svc.audioSources) != 0 {
 			svc.logger.Debug("not starting streams due to no stream config being set")
 		}
-		noopServer, err := gostream.NewStreamServer(streams...)
+		noopServer, err := webstream.NewServer(streams...)
 		return &StreamServer{noopServer, false}, err
 	}
 
@@ -209,7 +209,7 @@ func (svc *webService) makeStreamServer(ctx context.Context) (*StreamServer, err
 		streamTypes = append(streamTypes, false)
 	}
 
-	streamServer, err := gostream.NewStreamServer(streams...)
+	streamServer, err := webstream.NewServer(streams...)
 	if err != nil {
 		return nil, err
 	}
@@ -343,7 +343,7 @@ func (svc *webService) initStreamServer(ctx context.Context, options *weboptions
 	if err := svc.rpcServer.RegisterServiceServer(
 		ctx,
 		&streampb.StreamService_ServiceDesc,
-		svc.streamServer.Server.ServiceServer(),
+		svc.streamServer.Server,
 		streampb.RegisterStreamServiceHandlerFromEndpoint,
 	); err != nil {
 		return err
